Transfer LinkInspector settings from the 'OSGi console' to the 'Granit' console. #47 -fixed git comments
